+++ packages/editor/web/src/RichContentEditor/RichContentEditor.tsx
@@ -148,6 +148,8 @@ export interface RichContentEditorProps extends PartialDraftEditorProps {
   ) => DraftEditorProps['handleReturn'];
   tablePluginMenu?: boolean;
   callOnChangeOnNewEditorState?: boolean;
+  localeResource?: Record<string, string>;
+  maxTextLength?: number;
   /** This is a legacy API, chagnes should be made also in the new Ricos Editor API **/
 }
 
@@ -181,7 +183,7 @@ class RichContentEditor extends Component<RichContentEditorProps, State> {
   customStyleFn: DraftEditorProps['customStyleFn'];
   toolbars;
   innerRCECustomStyleFn;
-
+  getSelectedText: (editorState: EditorState) => string;
   static defaultProps: Partial<RichContentEditorProps> = {
     config: {},
     spellCheck: true,
@@ -240,6 +242,7 @@ class RichContentEditor extends Component<RichContentEditorProps, State> {
     this.copySource = registerCopySource(this.editor);
     preventWixFocusRingAccessibility(this.editorWrapper);
     this.reportDebuggingInfo();
+    this.preloadLibs();
     this.props.helpers?.onOpenEditorSuccess?.(Version.currentVersion);
   }
 
@@ -253,6 +256,15 @@ class RichContentEditor extends Component<RichContentEditorProps, State> {
     this.updateBounds = () => '';
     if (this.copySource) {
       this.copySource.unregister();
+    }
+  }
+
+  // imports dynamic chunks conditionally
+  preloadLibs() {
+    if (this.props.maxTextLength && this.props.maxTextLength > 0) {
+      import(
+        /* webpackChunkName: getSelectedText */ 'wix-rich-content-editor-common/libs/getSelectedText'
+      ).then(({ getSelectedText }) => (this.getSelectedText = getSelectedText));
     }
   }
 
@@ -584,14 +596,16 @@ class RichContentEditor extends Component<RichContentEditorProps, State> {
       }
     }
 
-    const { config, isInnerRCE } = this.props;
-    const resultEditorState = handlePastedText(
+    const { config, isInnerRCE, maxTextLength } = this.props;
+    const resultEditorState = handlePastedText({
       text,
       html,
       editorState,
-      isInnerRCE,
-      this.getHeadings(config)
-    );
+      pasteWithoutAtomic: isInnerRCE,
+      customHeadings: this.getHeadings(config),
+      maxTextLength,
+      getSelectedText: this.getSelectedText,
+    });
     this.updateEditorState(resultEditorState);
 
     return 'handled';
@@ -655,7 +669,7 @@ class RichContentEditor extends Component<RichContentEditorProps, State> {
   setInPluginEditingMode = (shouldEnable: boolean) => {
     // As explained in https://github.com/facebook/draft-js/blob/585af35c3a8c31fefb64bc884d4001faa96544d3/src/component/handlers/DraftEditorModes.js#L14
     const mode = shouldEnable ? 'render' : 'edit';
-    this.editor.setMode(mode);
+    this.editor?.setMode(mode);
     this.inPluginEditingMode = shouldEnable;
     const toolbarsToIgnore: ToolbarsToIgnore = shouldEnable ? ['SideToolbar'] : [];
     this.setState({ toolbarsToIgnore });
@@ -716,6 +730,20 @@ class RichContentEditor extends Component<RichContentEditorProps, State> {
       // in general, disables any input click on atomic blocks
       handled = 'handled';
     }
+
+    // input is ignored if length > maxTextLength (if maxTextLength is set)
+    if (this.props.maxTextLength && this.props.maxTextLength > 0) {
+      const contentLength = this.state.editorState.getCurrentContent().getPlainText('').length;
+      const selectedTextLength = this.getSelectedText(this.state.editorState).length;
+      if (contentLength - selectedTextLength > this.props.maxTextLength - 1) {
+        // eslint-disable-next-line no-console
+        console.debug(
+          `text editing prevented due to maxTextLength limitation (${this.props.maxTextLength})`
+        );
+        handled = 'handled';
+      }
+    }
+
     return handled || 'not-handled';
   };