+++ packages/plugin-table/web/src/components/CellRenderer.js
@@ -166,9 +166,8 @@ export default class Cell extends Component {
         ref={this.setTdRef}
         className={classNames(
           styles.cell,
-          shouldShowSelectedStyle && styles.selected,
-          range?.length === 1 && styles.singleSelection,
-          isContainedInHeader && styles.header
+          isContainedInHeader && styles.header,
+          shouldShowSelectedStyle && !isEditing && styles.selected
         )}
         onMouseDown={onMouseDown}
         onMouseOver={onMouseOver}
@@ -199,6 +198,15 @@ export default class Cell extends Component {
           </Editor>
         </div>
         <CellBorders borders={cellBorders} />
+        <div
+          style={{
+            height: shouldShowSelectedStyle ? this.tdRef?.offsetHeight : 0,
+          }}
+          className={classNames(
+            shouldShowSelectedStyle && styles.selected,
+            range?.length === 1 && styles.singleSelection
+          )}
+        />
       </Tag>
     );
   }