+++ packages/ricos-content/web/src/migrateSchema/fromDraft/fromDraft.ts
@@ -1,14 +1,14 @@
 /* eslint-disable no-console, fp/no-loops, no-case-declarations */
 
 import { cloneDeep, isEmpty } from 'lodash';
-import { RicosContent as RicosContentDraft, RicosContentBlock } from '../..';
-import { BlockType, FROM_DRAFT_LIST_TYPE, HeaderLevel, NodeType } from '../consts';
-import { RicosContent, RicosNode, google } from 'ricos-schema';
+import { RicosContent, RicosContentBlock } from '../..';
+import { BlockType, FROM_DRAFT_LIST_TYPE, HeaderLevel } from '../consts';
+import { rich_content } from 'ricos-schema';
 import { genKey } from '../generateRandomKey';
 import { getTextNodes } from './getTextNodes';
 import { getEntity, parseBlockData } from './getRicosEntityData';
 
-const createTimestamp = (): google.protobuf.Timestamp => {
+const createTimestamp = (): rich_content.Metadata['lastEdited'] => {
   const timeMS = Date.now();
   return {
     seconds: Math.floor(timeMS / 1000),
@@ -16,12 +16,12 @@ const createTimestamp = (): google.protobuf.Timestamp => {
   };
 };
 
-export const ensureRicosContent = (content: RicosContent | RicosContentDraft) =>
+export const ensureRicosContent = (content: rich_content.RichContent | RicosContent) =>
   'blocks' in content ? fromDraft(content) : content;
 
-export const fromDraft = (draftJSON: RicosContentDraft): RicosContent => {
+export const fromDraft = (draftJSON: RicosContent): rich_content.RichContent => {
   const { blocks, entityMap, VERSION: version } = cloneDeep(draftJSON);
-  const nodes: RicosNode[] = [];
+  const nodes: rich_content.Node[] = [];
 
   const parseBlocks = (index = 0) => {
     const block = blocks[index];
@@ -65,7 +65,7 @@ export const fromDraft = (draftJSON: RicosContentDraft): RicosContent => {
     }
   };
 
-  const parseAtomicBlock = (block: RicosContentBlock): RicosNode => {
+  const parseAtomicBlock = (block: RicosContentBlock): rich_content.Node => {
     return {
       key: block.key,
       nodes: [],
@@ -73,22 +73,22 @@ export const fromDraft = (draftJSON: RicosContentDraft): RicosContent => {
     };
   };
 
-  const parseQuoteBlock = (block: RicosContentBlock): RicosNode => ({
+  const parseQuoteBlock = (block: RicosContentBlock): rich_content.Node => ({
     key: block.key,
-    type: NodeType.Blockquote,
+    type: rich_content.Node.Type.BLOCKQUOTE,
     nodes: [parseTextBlock(block)],
   });
 
-  const parseCodeBlock = (block: RicosContentBlock): RicosNode => ({
+  const parseCodeBlock = (block: RicosContentBlock): rich_content.Node => ({
     key: block.key,
-    type: NodeType.CodeBlock,
+    type: rich_content.Node.Type.CODEBLOCK,
     nodes: getTextNodes(block, entityMap),
-    ricosCode: {
+    codeData: {
       ...parseBlockData(block.data),
     },
   });
 
-  const parseHeadingBlock = (block: RicosContentBlock): RicosNode => {
+  const parseHeadingBlock = (block: RicosContentBlock): rich_content.Node => {
     const getLevel = (blockType: string) => {
       if (Object.keys(HeaderLevel).includes(blockType)) {
         return HeaderLevel[blockType];
@@ -98,8 +98,8 @@ export const fromDraft = (draftJSON: RicosContentDraft): RicosContent => {
     };
     return {
       key: block.key,
-      type: NodeType.Heading,
-      ricosHeading: {
+      type: rich_content.Node.Type.HEADING,
+      headingData: {
         level: getLevel(block.type),
         depth: block.depth || undefined,
         ...parseBlockData(block.data),
@@ -108,11 +108,11 @@ export const fromDraft = (draftJSON: RicosContentDraft): RicosContent => {
     };
   };
 
-  const parseTextBlock = (block: RicosContentBlock): RicosNode => {
-    const textWrapperNode: RicosNode = {
+  const parseTextBlock = (block: RicosContentBlock): rich_content.Node => {
+    const textWrapperNode: rich_content.Node = {
       key: genKey(),
-      type: NodeType.Paragraph,
-      ricosParagraph: {
+      type: rich_content.Node.Type.PARAGRAPH,
+      paragraphData: {
         ...parseBlockData(block.data),
       },
       nodes: [],
@@ -125,8 +125,8 @@ export const fromDraft = (draftJSON: RicosContentDraft): RicosContent => {
       case BlockType.Blockquote:
       case BlockType.OrderedListItem:
       case BlockType.UnorderedListItem:
-        textWrapperNode.ricosParagraph = {
-          ...textWrapperNode.ricosParagraph,
+        textWrapperNode.paragraphData = {
+          ...textWrapperNode.paragraphData,
           depth: block.depth,
         };
         break;
@@ -142,17 +142,20 @@ export const fromDraft = (draftJSON: RicosContentDraft): RicosContent => {
     return textWrapperNode;
   };
 
-  const createListItem = (block: RicosContentBlock): RicosNode => ({
+  const createListItem = (block: RicosContentBlock): rich_content.Node => ({
     key: block.key,
-    type: NodeType.ListItem,
+    type: rich_content.Node.Type.LIST_ITEM,
     nodes: [parseTextBlock(block)],
   });
 
-  const isListBlock = (block: RicosContentBlock): boolean => !!FROM_DRAFT_LIST_TYPE[block.type];
+  const isListBlock = (block: RicosContentBlock): boolean =>
+    FROM_DRAFT_LIST_TYPE[block.type] !== undefined;
 
-  const parseListBlocks = (listStartIndex: number): { node: RicosNode; nextIndex: number } => {
+  const parseListBlocks = (
+    listStartIndex: number
+  ): { node: rich_content.Node; nextIndex: number } => {
     const { type: listType, depth } = blocks[listStartIndex];
-    const listNodes: RicosNode[] = [];
+    const listNodes: rich_content.Node[] = [];
     let searchIndex = listStartIndex;
     let nextBlock = blocks[searchIndex];
 
@@ -180,29 +183,28 @@ export const fromDraft = (draftJSON: RicosContentDraft): RicosContent => {
 
   parseBlocks();
 
-  const ricosContentMessage = RicosContent.fromObject({
-    doc: {
-      nodes,
+  const ricosContentMessage = rich_content.RichContent.fromObject({
+    nodes,
+    metadata: {
       lastEdited: createTimestamp(),
+      updatedVersion: version || '',
+      createdVersion: version || '',
     },
-    selection: {
-      anchorNode: nodes[0].key,
-    },
-    version: version || '',
   });
 
-  const err = RicosContent.verify(ricosContentMessage);
+  const err = rich_content.RichContent.verify(ricosContentMessage);
   if (err) {
     console.log('ERROR! Invalid content');
     console.log(err);
     process.exit(1);
   }
 
-  const ricosContent = RicosContent.toObject(ricosContentMessage, {
+  const ricosContent = rich_content.RichContent.toObject(ricosContentMessage, {
     arrays: true,
     enums: String,
     longs: Number,
   });
 
-  return ricosContent;
+  // Should consider creating a new type for RichContent with string enums
+  return ricosContent as rich_content.RichContent;
 };