+++ CHANGELOG.md
@@ -23,12 +23,47 @@
   - [#1939](https://github.com/wix-incubator/rich-content/pull/1931) allow users to toggle the expand mode option on the image-settings
   </summary>
 
+
+### :bug: Bug Fix
+- `unsupported-blocks`
+  - [#2058](https://github.com/wix/ricos/pull/2058) bg color update
+</details>
+<hr/>
+
+## 8.17.10 (Jan 27, 2021)
+### :bug: Bug Fix
+- `link`
+  - [#2049](https://github.com/wix/ricos/pull/2049) fix url data changing rel/target when only one of them is present
+
+## 8.17.9 (Jan 26, 2021)
+### :bug: Bug Fix
+- `link`
+  - [#2038](https://github.com/wix/ricos/pull/2038) fix anchor's tag overriding url,add `href` value & remove `siteUrl` prop
+
+## 8.17.8 (Jan 25, 2021)
+### :rocket: New Feature
+- `editor`
+  - [#2034](https://github.com/wix-incubator/rich-content/pull/2034) maxTextLength prop added
+### :bug: Bug Fix
+- `table`
+  - [#2046](https://github.com/wix/ricos/pull/2046) cell selection background color style
+
+## 8.17.7 (Jan 25, 2021)
+### :bug: Bug Fix
+- `ricos-common`
+  - [#2041](https://github.com/wix/ricos/pull/2041) remove draft dependency in migration tool
+
+## 8.17.6 (Jan 22, 2021)
 ### :rocket: New Feature
 - `ricos-content`
   - [#2006](https://github.com/wix-incubator/rich-content/pull/2006) `toPlainText` utility converts rich content to plain text
-
-</details>
-<hr/>
+### :house: Internal
+- `example`
+  - [#2026](https://github.com/wix/ricos/pull/2026) convert to TypeScript
+- `storybook`
+  - [#2026](https://github.com/wix/ricos/pull/2026) convert to TypeScript
+- `e2e`
+  - [#2026](https://github.com/wix/ricos/pull/2026) convert test-env to TypeScript
 
 ## 8.17.5 (Jan 20, 2021)
 ### :house: Internal